refactor push-notifications
extracted separate service to create notifications and moved everything related to pubnub into its own package
