+++ src/main/java/com/inovex/zabbixmobile/OnUpgradeReceiver.java
@@ -22,15 +22,15 @@ import android.content.Context;
 import android.content.Intent;
 import android.util.Log;
 
-import com.inovex.zabbixmobile.push.PushService;
+import com.inovex.zabbixmobile.push.pubnub.PubnubPushService;
 
 public class OnUpgradeReceiver extends BroadcastReceiver {
 
 	@Override
 	public void onReceive(Context context, Intent intent) {
 		Log.d("UpdateBroadcastReceiver", "onUpdate");
-		PushService.killPushService(context);
-		PushService.startOrStopPushService(context);
+		PubnubPushService.killPushService(context);
+		PubnubPushService.startOrStopPushService(context);
 	}
 
 }
